+++ src/shared/kakaoRedirect.js
@@ -8,10 +8,6 @@ const KakaoRedirect = props => {
 
   // 인가코드
   let code = new URL(window.location.href).searchParams.get("code");
-  //post방식
-  // const codeinfo = {
-  //   code: code,
-  // };
 
   React.useEffect(() => {
     dispatch(userCreators.kakaologinMiddleware(code));
